claudebot/requirements.txt
anthropic>=0.42.0
python-telegram-bot>=22.0
discord.py>=2.4.0
python-dotenv>=1.0.0
